+++ src/NamingService/Test/NamingServiceTest.cxx
@@ -1169,7 +1169,7 @@ void NamingServiceTest::_destroyDirectoryRecurs(std::string path)
   if (_NS.Change_Directory(path.c_str()))
     {
       std::vector<std::string> subdirs = _NS.list_subdirs();
-      for (int i=0; i<subdirs.size(); i++)
+      for (size_t i=0; i<subdirs.size(); i++)
         {
           std::string subpath=path + "/" +subdirs[i];
           _destroyDirectoryRecurs(subpath);
